made the pie chart larger
