+++ client/components/D3/RenderActiveShape.js
@@ -18,10 +18,10 @@ export const RenderActiveShape = props => {
   } = props
   const sin = Math.sin(-RADIAN * midAngle)
   const cos = Math.cos(-RADIAN * midAngle)
-  const sx = cx + (outerRadius + 10) * cos
-  const sy = cy + (outerRadius + 10) * sin
-  const mx = cx + (outerRadius + 30) * cos
-  const my = cy + (outerRadius + 30) * sin
+  const sx = cx + (outerRadius + 15) * cos
+  const sy = cy + (outerRadius + 15) * sin
+  const mx = cx + (outerRadius + 35) * cos
+  const my = cy + (outerRadius + 35) * sin
   const ex = mx + (cos >= 0 ? 1 : -1) * 22
   const ey = my
   const textAnchor = cos >= 0 ? 'start' : 'end'
