+++ client/components/DonationDialog.js
@@ -25,9 +25,11 @@ class DonationDialogSlide extends React.Component {
   render () {
     return (
       <div>
-        
+
       </div>
     )
-    
+
   }
 }
+
+export default DonationDialogSlide
